pyproject.toml
[tool.poetry]
name = "badbot"
version = "0.1.0"
description = ""
authors = ["Ilya Polotsky <ipolo.box@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.5"
emoji = "^1.2.0"
Pillow = "^8.2.0"
PyQRCode = "^1.2.1"
pypng = "^0.0.20"
firebase-admin = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
